Fix potential buffer overflow receiving messages
A slave device could indicate that there are 255 bytes in the response
here:
  u8BytesLeft = u8ModbusADU[2];

There are only 256 bytes of space in ModbusADU so the additional 255
bytes would overflow the buffer.

Increase the size of ModbusADUSize so that it is large enough to
handle the byte at index 255 without wrapping around to 0, and use it
to check that there is enough space for the response. When we run out
of space, return a new error type ku8MBResponseTooLarge (0xE4).

(cherry picked from commit 1df4097f28ea84142b07cf47831779913ddcf9f0)